--- sentences.txt
@@ -1,25 +1,0 @@
-I ate lunch with my friends.
-He told me about his trip to Germany.
-Every Go program is made up of packages.
-I can hold my breath underwater for two minutes!
-You can add your own writing prompts. Check out sentences.txt!
-Is math related to science?
-Where did the time go? We may never know.
-Brad invited me to the bar, but I declined.
-Someday, I would like to visit Indonesia.
-My face turned to stone when I heard the news.
-I recieved a letter in the mail.
-My neighbor knocked on my door last night.
-The quick brown fox jumps over the lazy dog.
-Creating writing prompts is harder than I thought it would be!
-What is the difference between encoding and marshalling?
-A switch statement is a shorter way to write a sequence of if-else statements.
-What is your favorite color?
-The weather was pleasant, so I decided to take a walk.
-I forgot my umbrella today, so I hope it doesn't rain.
-It was me, Dio!
-Watch those wrist rockets!
-I listen to music very often.
-We all agreed to meet in the conference room and review the test material.
-How many more writing prompts should I write?
-Nothing is worse than a stressful weekend.
